+++ src/App.js
@@ -1,10 +1,11 @@
 import React, { Component } from "react";
-import Header from './Header';
+import Header from './Components/Header';
 
 class App extends Component {
   render() {
     return (
       <Header />
+      <p
     );
   }
 }
